Updating test_all.sh to cover all cases in check_all.sh.
This commit will update test_all.sh and also remove the ability
from test_all.sh to check the output, since this is now performed
in check_all.sh. We want to make test_all.sh check only the run
time and not the output because writing the output to the disk
may consume a nontrivial amount of time for some benchmarks, such
as basicmath.