[artifact17] Remove run.sh from build in Dockerfile
I still could not figure out why run.sh is not working when Dockerfile
tries to run it directly. Oddly, it does work after image is finished
and the user starts a new container. The problem seems to be with perf
report. This commit changes instructions in Dockerfile to tell the user
to execute run.sh by herself.